# docs/元数据服务/openapi.json
{
  "components": {
    "schemas": {
...
      "_labels_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "properties": {
                "name": {
                  "format": "string",
                  "description": "名称 label should contain alphanumeric characters, _ or -",
                  "type": "string"
                },
                "description": {
                  "format": "string",
                  "description": "描述",
                  "type": "string"
                },
                "id": {
                  "format": "int64",
                  "description": "No comments found.",
                  "type": "integer"
                },
                "gmtCreate": {
                  "format": "string",
                  "description": "创建时间",
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "_labels_bind_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
...
      "_labels__ids_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
      "_labels_bind_bulkResponse": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
      "_labels_bind_Request": {
        "properties": {
          "guid": {
            "format": "string",
            "description": "元数据标识",
            "type": "string"
          },
          "labels": {
            "format": "string",
            "description": "元数据标签",
            "type": "string"
          }
        },
        "required": [
          "guid"
        ]
      },
...
      "_classify_bind__guid_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "guid": {
                "format": "string",
                "description": "元数据标识",
                "type": "string"
              },
              "classify": {
                "format": "string",
                "description": "分类",
                "type": "string"
              }
            }
          }
        }
      },
      "_labels__id_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
...
      "_labels_Request": {},
...
      "_labels_pageResponse": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "total": {
                "format": "int64",
                "description": "数据总量",
                "type": "integer"
              },
              "pageNum": {
                "format": "int32",
                "description": "当前页",
                "type": "integer"
              },
              "pageSize": {
                "format": "int32",
                "description": "每页大小",
                "type": "integer"
              },
              "list": {
                "description": "返回数据data",
                "type": "array",
                "items": {
                  "properties": {
                    "name": {
                      "format": "string",
                      "description": "名称 label should contain alphanumeric characters, _ or -",
                      "type": "string"
                    },
                    "description": {
                      "format": "string",
                      "description": "描述",
                      "type": "string"
                    },
                    "id": {
                      "format": "int64",
                      "description": "No comments found.",
                      "type": "integer"
                    },
                    "gmtCreate": {
                      "format": "string",
                      "description": "创建时间",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "_labels_bind__guid_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "guid": {
                "format": "string",
                "description": "元数据标识",
                "type": "string"
              },
              "labels": {
                "format": "string",
                "description": "元数据标签",
                "type": "string"
              }
            }
          }
        }
      },
...
      "_labels_bind_bulkRequest": {
        "properties": {
          "guids": {
            "description": "元数据标识",
            "type": "array",
            "items": {}
          },
          "labels": {
            "format": "string",
            "description": "元数据标签",
            "type": "string"
          }
        },
        "required": [
          "guids",
          "labels"
        ]
      },
...
      "_labels__id_Request": {
        "properties": {
          "name": {
            "format": "string",
            "description": "名称 label should contain alphanumeric characters, _ or -",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "描述",
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  },
  "servers": [
    {
      "url": "http://172.20.0.13:8782/mtd"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/classify/bind/{guid}": {
      "get": {
        "summary": "查询元数据分类绑定关系",
        "deprecated": false,
        "description": "查询元数据分类绑定关系",
        "tags": [
          "MtdClassifyAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify_bind__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"vjo0cx\",\n  \"tips\": \"5gnt6e\",\n  \"data\": {\n    \"guid\": \"59\",\n    \"classify\": \"dztsfl\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据guid",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/labels/{id}": {
      "put": {
        "summary": "修改元数据标签",
        "deprecated": false,
        "description": "修改元数据标签",
        "tags": [
          "MtdLabelsController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_labels__id_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"name\": \"rochel.emard\",\n  \"description\": \"zy1iem\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels__id_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"v5g2y3\",\n  \"tips\": \"57t92i\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "update",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "example": "721",
            "name": "id"
          }
        ]
      }
    },
...
    "/labels/{ids}": {
      "delete": {
        "summary": "删除元数据标签",
        "deprecated": false,
        "description": "删除元数据标签",
        "tags": [
          "MtdLabelsController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels__ids_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"jjtybq\",\n  \"tips\": \"zt5p1a\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "delete",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "id字符串",
            "required": true,
            "example": "ws95zi",
            "name": "ids"
          }
        ]
      }
    },
    "/labels/page": {
      "get": {
        "summary": "分页查询元数据标签",
        "deprecated": false,
        "description": "分页查询元数据标签",
        "tags": [
          "MtdLabelsController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_pageResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"zrx86z\",\n  \"tips\": \"9vqk2u\",\n  \"data\": {\n    \"total\": 272,\n    \"pageNum\": 14,\n    \"pageSize\": 10,\n    \"list\": [\n      {\n        \"name\": \"rochel.emard\",\n        \"description\": \"mzl69c\",\n        \"id\": 362,\n        \"gmtCreate\": \"2021-10-25 16:17:38\"\n      }\n    ]\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "listByPage",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "开始时间",
            "required": false,
            "example": "cznt9d",
            "name": "beginDay"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "结束时间",
            "required": false,
            "example": "5sortv",
            "name": "endDay"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "名称",
            "required": false,
            "example": "rochel.emard",
            "name": "name"
          }
        ]
      }
    },
...
    "/labels/bind/bulk": {
      "post": {
        "summary": "批量绑定元数据标签",
        "deprecated": false,
        "description": "批量绑定元数据标签",
        "tags": [
          "MtdLabelsAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_labels_bind_bulkRequest"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"guids\": [\n    \"6ocbxe\"\n  ],\n  \"labels\": \"hk6uph\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_bind_bulkResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"x1pdy8\",\n  \"tips\": \"utn2c3\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "bulk",
        "parameters": []
      }
    },
...
    "/labels/bind/{guid}": {
      "get": {
        "summary": "查询元数据标签绑定关系",
        "deprecated": false,
        "description": "查询元数据标签绑定关系",
        "tags": [
          "MtdLabelsAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_bind__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"rs441k\",\n  \"tips\": \"fs8cw7\",\n  \"data\": {\n    \"guid\": \"59\",\n    \"labels\": \"pkfgtc\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据唯一标识",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/labels/": {
      "post": {
        "summary": "新增元数据标签",
        "deprecated": false,
        "description": "新增元数据标签",
        "tags": [
          "MtdLabelsController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_labels_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"name\": \"rochel.emard\",\n  \"description\": \"3xy241\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"kg8vbd\",\n  \"tips\": \"ddxxbr\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "insert",
        "parameters": []
      },
      "get": {
        "summary": "查询元数据标签",
        "deprecated": false,
        "description": "查询元数据标签",
        "tags": [
          "MtdLabelsController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"upq1jg\",\n  \"tips\": \"iwdgcm\",\n  \"data\": [\n    {\n      \"name\": \"rochel.emard\",\n      \"description\": \"bm9kip\",\n      \"id\": 928,\n      \"gmtCreate\": \"2021-10-25 16:17:38\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "listByAll",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "名称 label should contain alphanumeric characters, _ or -",
            "required": true,
            "example": "rochel.emard",
            "name": "name"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "描述",
            "required": false,
            "example": "ssuqq5",
            "name": "description"
          }
        ]
      }
    },
...
    "/labels/bind/": {
      "post": {
        "summary": "新增元数据标签绑定关系",
        "deprecated": false,
        "description": "新增元数据标签绑定关系",
        "tags": [
          "MtdLabelsAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_labels_bind_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"guid\": \"59\",\n  \"labels\": \"rl8c35\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_bind_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"llf497\",\n  \"tips\": \"hxffo2\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "insert",
        "parameters": []
      },
      "put": {
        "summary": "修改元数据标签绑定关系",
        "deprecated": false,
        "description": "修改元数据标签绑定关系",
        "tags": [
          "MtdLabelsAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_labels_bind_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"guid\": \"59\",\n  \"labels\": \"bt7wr3\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_bind_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"1p8s2u\",\n  \"tips\": \"y9r9se\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "update",
        "parameters": []
      }
    },
...
  },
  "info": {
    "title": "dm-metadata",
    "version": "1.0.0"
  }
}